// src/module/zone/interfaces/detail.interface.ts
export interface IDetail {
  // 门牌楼属性
  readonly MLPSXDM: string;
  // 末端地址标志
  readonly MDJD: string;
  // 号座
  readonly HAOZUO: string;
  // GUID编码
  readonly SYSTEMID: string;
  // 地市编码
  readonly DSBM: string;
  // 地址类型
  readonly DZLX: string;
  // 地址全称
  readonly DZMC: string;
  // 门牌号码
  readonly MPH: string;
  // 小区名称
  readonly XQMC: string;
  // 小区编号
  readonly XQDM: string;
  // 楼房名称
  readonly JZWMC: string;
  // 楼房编号
  readonly JZWDM: string;
  // 房号
  readonly DYFH: string;
  // 数据归属单位代码
  readonly GAJGJGDM: string;
  // 数据归属单位名称
  readonly GAJGJGMC: string;
  // 警务区编号
  readonly JWWGDM: string;
  // 警务区名称
  readonly JWWGMC: string;
  // 乡镇（街道）代码
  readonly XZJDDM: string;
  // 乡镇街道
  readonly XZJDMC: string;
  // 区乡村编号
  readonly SQJCWHDM: string;
  // 区乡村
  readonly SQJCWHMC: string;
  // 地名编号
  readonly DMDM: string;
  // 地名
  readonly DMMC: string;
  // 地（住）址存在标识
  readonly DZZCZBS: string;
  // 地（住）址在用标识
  readonly DZZZYBS: string;
  // 启用日期
  readonly QYRQ: string;
  // 停用日期 
  readonly TYRQ: string;
  // X坐标
  readonly MAPX: string;
  // X坐标
  readonly MAPY: string;
  // 登记时间
  readonly DJSJ: string;
  // 登记人姓名
  readonly DJR_XM: string;
  // 登记人单位代码
  readonly DJDW_GAJGJGDM: string;
  // 登记人单位名称
  readonly DJDW_GAJGJGMC: string;
  // 单元梯位代码
  readonly CELL_ID: string;
  // 单元
  readonly CELL: string;
  // 地址元素类型
  readonly DZYSLX: string;
  // 区县名称
  readonly QU: string;
  // 省市县区编码
  readonly QU_ID: string;
  // 属性
  readonly PROPERTY: string;
}